Minor fixes

+++ RestApi/DTO/Temperature/AddTemperatureDTO.cs
@@ -1,13 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace rest_api.DTO.Temperature
 {
     public class AddTemperatureDTO
     {
         public int Id { get; set; }
-        public int Temp { get; set; }
+        public double Temp { get; set; }
     }
 }
