+++ RestApi/Models/Temperature.cs
@@ -3,7 +3,7 @@ namespace rest_api.Models
     public class Temperature
     {
         public int Id { get; set; }
-        public int Temp { get; set; }
+        public double Temp { get; set; }
         public DateTime DateTime { get; set; }
     }
 }
